Move message providers into separate module
